fix: version bump up (#8621)
