+++ src/modules/70-pipeline/pages/pipelines/PipelinesPage.tsx
@@ -83,13 +83,13 @@ import { useFeatureFlag, useFeatureFlags } from '@common/hooks/useFeatureFlag'
 import { FeatureFlag } from '@common/featureFlags'
 import useRBACError from '@rbac/utils/useRBACError/useRBACError'
 import { NewPipelinePopover } from '@pipeline/pages/pipelines/views/NewPipelinePopover/NewPipelinePopover'
+import { deploymentTypeLabel } from '@pipeline/utils/DeploymentTypeUtils'
 import { PipelineGridView } from './views/PipelineGridView'
 import { PipelineListView } from './views/PipelineListView'
 import PipelineFilterForm from '../pipeline-deployment-list/PipelineFilterForm/PipelineFilterForm'
 import pipelineIllustration from './images/deploypipeline-illustration.svg'
 import buildpipelineIllustration from './images/buildpipeline-illustration.svg'
 import flagpipelineIllustration from './images/flagpipeline-illustration.svg'
-import { deploymentTypeLabel } from './PipelineListUtils'
 import css from './PipelinesPage.module.scss'
 
 export enum Sort {
